Deleted the legacysite and moved jekyllsite to server root
After this commit, the repository will be the result of `rm -rf`-ing the
legacysite and dragging the jekyll site code into the server root and
deleting the now-empty folder it used to all be in.

I have a backup of the legacy site on my local machine, so don't worry.